feat: add Redis generation handoff continuation
