+++ crates/simulator-core/src/broadcaster/redis_streams.rs
@@ -44,7 +44,7 @@ pub struct BroadcasterRedisStreamEntry {
     pub payload_json: String,
 }
 
-/// Redis replay cursor captured with an HTTP snapshot session.
+/// Redis replay checkpoint captured with an HTTP snapshot session.
 #[derive(Debug, Clone, PartialEq, Eq, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct BroadcasterRedisReplayBoundary {
@@ -613,8 +613,9 @@ mod tests {
     };
     use crate::broadcaster::{
         BroadcasterBackend, BroadcasterBackendHead, BroadcasterContractError, BroadcasterEnvelope,
-        BroadcasterHeartbeat, BroadcasterMessageKind, BroadcasterPayload, BroadcasterStateEntry,
-        BroadcasterUpdateMessage, BroadcasterUpdatePartition,
+        BroadcasterGenerationHandoff, BroadcasterHeartbeat, BroadcasterMessageKind,
+        BroadcasterPayload, BroadcasterProgress, BroadcasterStateEntry, BroadcasterUpdateMessage,
+        BroadcasterUpdatePartition,
     };
 
     #[test]
@@ -932,7 +933,163 @@ mod tests {
     }
 
     #[test]
-    fn redis_replay_boundary_uses_camel_case_shape_and_deterministic_cursor() -> Result<()> {
+    fn redis_stream_entry_omits_handoff_for_normal_progress() -> Result<()> {
+        let envelope = progress_envelope("stream-1", 8453, "snapshot-1", 6)?;
+        let entry = redis_entry(&envelope)?;
+
+        assert_eq!(entry.kind, BroadcasterMessageKind::Progress);
+        assert_eq!(entry.snapshot_id, Some("snapshot-1".to_string()));
+        assert_eq!(entry.backend_scope, "native,vm");
+        assert_eq!(entry.block_number, None);
+        assert_eq!(entry.observed_timestamp_ms, None);
+
+        let payload: serde_json::Value = serde_json::from_str(&entry.payload_json)?;
+        assert!(payload.get("handoff").is_none());
+
+        let value = serde_json::to_value(&entry)?;
+        assert!(value.get("block_number").is_none());
+        assert!(value.get("observed_timestamp_ms").is_none());
+
+        let decoded: BroadcasterRedisStreamEntry = serde_json::from_value(value)?;
+        assert_eq!(decoded, entry);
+
+        Ok(())
+    }
+
+    #[test]
+    fn redis_stream_entry_round_trips_progress_with_handoff() -> Result<()> {
+        let handoff = generation_handoff()?;
+        let progress = BroadcasterProgress::new_with_handoff(
+            8453,
+            "snapshot-2",
+            vec![BroadcasterBackend::Vm, BroadcasterBackend::Native],
+            "generation_reset",
+            handoff.clone(),
+        )?;
+        let envelope =
+            BroadcasterEnvelope::new("stream-2", 9, BroadcasterPayload::Progress(progress));
+        let entry = redis_entry(&envelope)?;
+
+        assert_eq!(entry.kind, BroadcasterMessageKind::Progress);
+        assert_eq!(entry.snapshot_id, Some("snapshot-2".to_string()));
+        assert_eq!(entry.backend_scope, "native,vm");
+
+        let payload: serde_json::Value = serde_json::from_str(&entry.payload_json)?;
+        assert_eq!(
+            payload["handoff"],
+            serde_json::json!({
+                "previousStreamId": "stream-1",
+                "previousEntryId": "7-42",
+                "baseHeads": [
+                    {"backend": "native", "blockNumber": 124},
+                    {"backend": "vm", "blockNumber": 125}
+                ]
+            })
+        );
+
+        let decoded_entry: BroadcasterRedisStreamEntry =
+            serde_json::from_value(serde_json::to_value(&entry)?)?;
+        let decoded_envelope: BroadcasterEnvelope =
+            serde_json::from_str(&decoded_entry.payload_json)?;
+        let BroadcasterPayload::Progress(decoded_progress) = decoded_envelope.payload else {
+            return Err(anyhow!("decoded payload should remain progress"));
+        };
+        assert_eq!(decoded_progress.handoff, Some(handoff));
+
+        Ok(())
+    }
+
+    #[test]
+    fn generation_handoff_rejects_empty_ids_and_duplicate_base_heads() -> Result<()> {
+        let error = BroadcasterGenerationHandoff::new("", "7-42", base_heads())
+            .err()
+            .ok_or_else(|| anyhow!("empty previous stream id should fail handoff construction"))?;
+        assert_eq!(
+            error,
+            BroadcasterContractError::RedisEntryEmptyField {
+                field: "handoff.previousStreamId",
+            }
+        );
+
+        let error = BroadcasterGenerationHandoff::new("stream-1", " ", base_heads())
+            .err()
+            .ok_or_else(|| anyhow!("empty previous entry id should fail handoff construction"))?;
+        assert_eq!(
+            error,
+            BroadcasterContractError::RedisEntryEmptyField {
+                field: "handoff.previousEntryId",
+            }
+        );
+
+        let duplicate_heads = vec![
+            BroadcasterBackendHead::new(BroadcasterBackend::Native, 124),
+            BroadcasterBackendHead::new(BroadcasterBackend::Native, 125),
+        ];
+        let error = BroadcasterGenerationHandoff::new("stream-1", "7-42", duplicate_heads)
+            .err()
+            .ok_or_else(|| anyhow!("duplicate handoff base heads should fail construction"))?;
+        assert_eq!(
+            error,
+            BroadcasterContractError::DuplicateBackendEntry {
+                context: "progress.handoff.baseHeads",
+                backend: BroadcasterBackend::Native,
+            }
+        );
+
+        Ok(())
+    }
+
+    #[test]
+    fn generation_handoff_deserialization_rejects_empty_ids_and_duplicate_base_heads() -> Result<()>
+    {
+        let empty_stream_error = handoff_decode_error(
+            serde_json::json!({
+                "previousStreamId": "",
+                "previousEntryId": "7-42",
+                "baseHeads": [
+                    {"backend": "native", "blockNumber": 124}
+                ]
+            }),
+            "empty previousStreamId",
+        )?;
+        assert!(empty_stream_error
+            .to_string()
+            .contains("handoff.previousStreamId must not be empty"));
+
+        let empty_entry_error = handoff_decode_error(
+            serde_json::json!({
+                "previousStreamId": "stream-1",
+                "previousEntryId": "",
+                "baseHeads": [
+                    {"backend": "native", "blockNumber": 124}
+                ]
+            }),
+            "empty previousEntryId",
+        )?;
+        assert!(empty_entry_error
+            .to_string()
+            .contains("handoff.previousEntryId must not be empty"));
+
+        let duplicate_heads_error = handoff_decode_error(
+            serde_json::json!({
+                "previousStreamId": "stream-1",
+                "previousEntryId": "7-42",
+                "baseHeads": [
+                    {"backend": "native", "blockNumber": 124},
+                    {"backend": "native", "blockNumber": 125}
+                ]
+            }),
+            "duplicate baseHeads",
+        )?;
+        assert!(duplicate_heads_error
+            .to_string()
+            .contains("duplicate backend entry Native"));
+
+        Ok(())
+    }
+
+    #[test]
+    fn redis_replay_boundary_uses_camel_case_shape_and_deterministic_checkpoint() -> Result<()> {
         let boundary = BroadcasterRedisReplayBoundary::new(
             "dsolver:broadcaster:prod-base:8453:events",
             "chain-8453-stream-7",
@@ -1044,6 +1201,44 @@ mod tests {
                 )],
             )?),
         ))
+    }
+
+    fn progress_envelope(
+        stream_id: &str,
+        chain_id: u64,
+        snapshot_id: &str,
+        message_seq: u64,
+    ) -> Result<BroadcasterEnvelope> {
+        Ok(BroadcasterEnvelope::new(
+            stream_id,
+            message_seq,
+            BroadcasterPayload::Progress(BroadcasterProgress::new(
+                chain_id,
+                snapshot_id,
+                vec![BroadcasterBackend::Vm, BroadcasterBackend::Native],
+                "generation_reset",
+            )?),
+        ))
+    }
+
+    fn generation_handoff() -> Result<BroadcasterGenerationHandoff, BroadcasterContractError> {
+        BroadcasterGenerationHandoff::new("stream-1", "7-42", base_heads())
+    }
+
+    fn base_heads() -> Vec<BroadcasterBackendHead> {
+        vec![
+            BroadcasterBackendHead::new(BroadcasterBackend::Vm, 125),
+            BroadcasterBackendHead::new(BroadcasterBackend::Native, 124),
+        ]
+    }
+
+    fn handoff_decode_error(
+        value: serde_json::Value,
+        context: &'static str,
+    ) -> Result<serde_json::Error> {
+        serde_json::from_value::<BroadcasterGenerationHandoff>(value)
+            .err()
+            .ok_or_else(|| anyhow!("{context} should fail"))
     }
 
     fn mixed_backend_update_envelope(
